fix HashMap so it no longer hashes all functions using their source string
This was causing different function instances created from the same source code
(e.g. as local functions inside the same function) to be hashed the same.

One of the effects was that when using mocks.module() and passing it an object
like so:
  mocks.module({myService: 42});
worked if it was the first such mocks.module() call for a single test. Any later
later such calls (whether passing them same or different object) were simply
ignored by the constructed mocks injector. This is so because all such passed
objects are actually represented as functions in an internal pongular module list
and, since they are all constructed using the same source, they are all hashed
the same so when pongular asks 'do I already have this module loaded?' it thinks
the answer is yes as it can find that same hash in its HashMap of already loaded
modules.

Matching workaround has long since been applied in angular sources.